lapcounter-server-lapdata/requirements.txt
aiomqtt==1.2.1
paho-mqtt==1.6.1
